refactor: add components/home folder
